% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/precision_functions.R
\name{APE.Med}
\alias{APE.Med}
\title{Average Percent Error (APE) Estimation}
\usage{
APE.Med(x)
}
\arguments{
\item{x}{Numeric vector of independent age determinations from a single
fish.}
}
\value{
The APE of the vector of ages \code{x} from an individual fish.
}
\description{
\code{APE.Med} calculates the average percent error (APE) for multiple age
   determinations on the same fish.  This function use the median of the
   observed ages as the basis for APE calculation.
}
\examples{
#Example giving an estimate with an APE.Med of 0.0
APE.Med(c(1, 1, 1))

#Example using missing data
APE.Med(c(1, NA, 2))

#Another example with missing data
APE.Med(c(3, 5, 3, 4, NA, 6))

#Example showing how APE.Med can be used to calculate mean APE for many fish
w <- seq(1, 10, 1)
x <- seq(2, 11, 1)
y <- seq(3, 12, 1)
z <- seq(4, 13, 1)
df <- data.table::rbindlist(apply(na.exclude(RedPorgy[, c(1, 2)]), 1,
 APE.Med))
df
mean(data.frame(df)$APE)
# Example with RedPorgy data
data(RedPorgy)
df <- data.table::rbindlist(apply(na.exclude(RedPorgy[, c(1,2)]), 1,
 APE.Med))
df
mean(data.frame(df)$APE)
}
\seealso{
\code{\link{APE}} for APE calculation using the mean age as
   the basis APE calculation.

Other APE functions: 
\code{\link{APE}()}
}
\concept{APE functions}
