% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Francis_bomb_functions.R
\name{bias.plot.BR}
\alias{bias.plot.BR}
\title{Bomb Radiocarbon Plots}
\usage{
bias.plot.BR(bias.out, plottype)
}
\arguments{
\item{bias.out}{list containing output from function
\code{\link[FishyR]{calc.bias}}}

\item{plottype}{integer, with possible values of 1 and 2 that controls what,
type of plots is plotted:
\describe{
  \item{1}{the simulated h distribution is plotted, along with arrows
     showing the h values for different values of percentage bias and
     broken lines showing the 95\% confidence interval for h}
  \item{2}{a cumulative distribution of percentage bias is plotted, with
     broken lines showing the 95\% confidence interval for bias}
  }}
}
\value{
A plot.
}
\description{
\code{bias.plot.BR} plots output from function \code{\link[FishyR]{calc.bias}}
}
\examples{
data(snapper)
snapper.red <- snapper[snapper$C14year >= 1955 & snapper$C14year <= 1972, ]
data(bluenose)
bluenose.red <- bluenose[bluenose$C14year >= 1955 &
  bluenose$C14year <= 1972, ]
bluenose.red <- bluenose.red[bluenose.red$C14 >= -506 &
  bluenose.red$C14 <= 94.0, ]
out <- calc.bias(snapper.red, bluenose.red, nsim = 5000, plot.type = 0,
  x = "C14year", y = "C14", age = "age", cy = "catch.year",
  age.err = "age.err", s.age = "samp.age", y.err = "C14.err",
  bias.range = c(-50, 50), bias.step = 1)
bias.plot.BR(out, plottype = 1)
bias.plot.BR(out, plottype = 2)
}
\seealso{
Other Bomb Radiocarbon Analyses: 
\code{\link{calc.bias}()},
\code{\link{calc.h}()},
\code{\link{closest.pt}()},
\code{\link{get.h.from.bias}()},
\code{\link{get.nearest}()},
\code{\link{get.ref.line}()},
\code{\link{sim.hdist}()}
}
\concept{Bomb Radiocarbon Analyses}
