% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_sets.R
\docType{data}
\name{bluenose}
\alias{bluenose}
\title{Bluenose (Hyperoglyphe antarctica) Bomb Radiocarbon Data}
\format{
A data frame with 12 rows and 9 variables
\describe{
 \item{fish}{factor, variable identifying individual fish}
 \item{age}{integer, age estimate}
 \item{catch.date}{factor, catch date}
 \item{C14year}{integer, estimated eyar of birth based on age and capture
    year}
 \item{C14}{numeric, C-C14 estimate}
 \item{C14.err}{numeric, C-14 estimate error estimate}
 \item{age.err}{numeric, age estimate error estimate}
 \item{samp.age}{integer, age estimate of analyzed sample}
 \item{catch.year}{capture year of fish}}
 @source Francis, R.I.C., Campana, S.E., and H.L. Neil.  2010.  Validation
    of fish ageing methods should involve bias estimation rather than
    hypothesis testing: a proposed approach for bomb radiocarbon
    validations.  Can. J. Fish. Aquat. Sci. 67: 1398-1408.
}
\usage{
bluenose
}
\description{
A dataset containing Bluenose Bomb Radiocarbon data, as provided in Francis
  et al. 2010.
}
\keyword{datasets}
