% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot_functions.R
\name{Boot.Diag}
\alias{Boot.Diag}
\title{Calculate bootstrap diagnostics from relative abundance index results}
\usage{
Boot.Diag(Rows, Columns, By, Statistics = c("CV", "Var"))
}
\arguments{
\item{Rows}{TBD}

\item{Columns}{TBD}

\item{By}{TBD}

\item{Statistics}{TBD}
}
\value{
Data frame with five columns:
  \describe{
    \item{Group}{}
    \item{Points}{}
    \item{Factor}{}
    \item{Frame}{}
  }
}
\description{
\code{Boot.Diag} generates a data frame containing summary statistics for
  various number of bootstrap runs.  Intent is to inform when uncertainty
  estimates of a relative abundance index stabilize.
}
\seealso{
Other Index Bootstrap Functions: 
\code{\link{Stat.Calc}()},
\code{\link{ZI.Index.Boot}()}
}
\concept{Index Bootstrap Functions}
