% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_sets.R
\docType{data}
\name{BSB}
\alias{BSB}
\title{Black Sea Bass Life History Data}
\format{
A data frame with 33,297 rows and 13 variables
\describe{
 \item{PCGSS}{unique identifier identifying individual fish}
 \item{Year}{year of capture}
 \item{Month}{month of capture}
 \item{Latitude}{latitude, in decimal degrees, where captured}
 \item{Depth}{depth, in meters, where captured}
 \item{TL}{total length, in mm}
 \item{SL}{standard length, in mm}
 \item{WholeWt}{whole weight, in grams}
 \item{GonadWt}{gonad weight, in grams}
 \item{Sex}{sex}
 \item{Mat}{maturity stage}
 \item{Age}{biological age}
 \item{Year.Class}{year born}
}
}
\source{
South Carolina Deparment of Natural Resources, Coastal Finfish Section
 (ballengerj@dnr.sc.gov)
}
\usage{
BSB
}
\description{
A dataset containing black sea bass life history data, as collected by the
   South Carolina Department of Natural Resources, Marine Resources Research
   Institute, Coastal Finfish Section, Reef Fish Survey.
}
\keyword{datasets}
