% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Francis_bomb_functions.R
\name{calc.bias}
\alias{calc.bias}
\title{Calculate Relationship Between "h" and Bias}
\usage{
calc.bias(ref, test, nsim = 5000, plot.type = 0, x, y, age, cy, age.err,
  s.age, y.err, bias.step = 5, bias.range = NULL)
}
\arguments{
\item{ref, test}{data frames representing reference and test data sets,
respectively.  Each row corresponds to one ref (test) datum and the
columns represent data needed to perform the bomb carbon analyses.
One column should represent each of the following input data in each data
frame:
\describe{
  \item{age}{estimated age}
  \item{age.err}{standard deviation of estimated age}
  \item{s.age}{estimated age of the fish when the material taken from
     the otolith/aging structure for bomb carbon analysis was
     incorporated into the otolith (if this column is omitted it will
     be assumed that \code{s.age = 0} for all samples)}
  \item{cy}{year of sampling, i.e. catch year}
  \item{x}{inferred year in which material taken from the otolith/aging
     structure for bomb carbon analysis was incorporated into the
     otolith (A warning will be output if \code{x} is inconsistent with
     other variables; should be \code{x = cy - (age - s.age)})}
  \item{y}{bomb carbon value (this will generally either be in C14 or
     fraction modern values)}
  \item{y.err}{standard deviation of bomb carbon value}
  }
Note, the columns in both the \code{ref} and \code{test} data frames
representing each variable should be named the same.}

\item{nsim}{number of simulated data sets used in generating the
distribution of h values}

\item{plot.type}{type of plot to produce (if 0, no plot is made).  See
function \code{\link[FishyR]{plot.bias}} for available plot types}

\item{age, age.err, s.age, cy, x, y, y.err}{Character strings representing the
column names for each of the required elements (see above) in the ref and
test data frames.}

\item{bias.step, bias.range}{See \code{\link[FishyR]{get.h.from.bias}} for
description}
}
\description{
\code{calc.bias} calculates a 95% confidence interval for the percentage
   bias in zone-count ages given appropriate bomb-carbon data.
}
\examples{
data(snapper)
snapper.red <- snapper[snapper$C14year >= 1955 & snapper$C14year <= 1972, ]
data(bluenose)
bluenose.red <- bluenose[bluenose$C14year >= 1955 &
  bluenose$C14year <= 1972, ]
bluenose.red <- bluenose.red[bluenose.red$C14 >= -506 &
  bluenose.red$C14 <= 94.0, ]
calc.bias(snapper.red, bluenose.red, nsim = 5000, plot.type = 0,
  x = "C14year", y = "C14", age = "age", cy = "catch.year",
  age.err = "age.err", s.age = "samp.age", y.err = "C14.err")
calc.bias(snapper.red, bluenose.red, nsim = 5000, plot.type = 1,
  x = "C14year", y = "C14", age = "age", cy = "catch.year",
  age.err = "age.err", s.age = "samp.age", y.err = "C14.err")
calc.bias(snapper.red, bluenose.red, nsim = 5000, plot.type = 2,
  x = "C14year", y = "C14", age = "age", cy = "catch.year",
  age.err = "age.err", s.age = "samp.age", y.err = "C14.err")
}
\seealso{
Other Bomb Radiocarbon Analyses: \code{\link{calc.h}},
  \code{\link{closest.pt}}, \code{\link{get.h.from.bias}},
  \code{\link{get.nearest}}, \code{\link{get.ref.line}},
  \code{\link{plot.bias}}, \code{\link{sim.hdist}}
}

