% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Francis_bomb_functions.R
\name{calc.h}
\alias{calc.h}
\title{Calculate the Median Horizontal Distance Between Linear Spline and Test
  Points}
\usage{
calc.h(line, test, age.err, test.x, test.y)
}
\arguments{
\item{line}{a list, with components x and y (each an n-vector),
containing the series of n points defining the line}

\item{test}{dataframe representing test data set as descrivbed in fucntion
\code{\link[FishyR]{calc.bombcarbon.bias}}}

\item{age.err, test.x, test.y}{character strings representing the names of the
columns in \code{test} containing the aging error estimates, the
x-variable, and the y-variable, respectively.}
}
\value{
Returns a numeric value representing the median horizontal distance
   from the reference line give the test data set.
}
\description{
\code{calc.h} calculates the median horizontal distance between
   the linear spline line and the set of points defined by test, where the
   horizontal distance is measured in standard errors.  Points whose
   y-values lie outside the range of the y-values for the line are excluded
   from teh calculation (because their horizontal distance from the line is
   undefined).
}
\examples{
data(snapper)
snapper.red <- snapper[snapper$C14year >= 1955 & snapper$C14year <= 1972, ]
data(bluenose)
bluenose.red <- bluenose[bluenose$C14year >= 1955 &
  bluenose$C14year <= 1972, ]
bluenose.red <- bluenose.red[bluenose.red$C14 >= -506 &
  bluenose.red$C14 <= 94.0, ]
ref <- get.ref.line(snapper.red$C14year, snapper.red$C14)
calc.h(line = ref, test = bluenose.red, age.err = "age.err", test.x =
  "C14year", test.y = "C14")
}
\seealso{
Other Bomb Radiocarbon Analyses: \code{\link{calc.bias}},
  \code{\link{closest.pt}}, \code{\link{get.h.from.bias}},
  \code{\link{get.nearest}}, \code{\link{get.ref.line}},
  \code{\link{plot.bias}}, \code{\link{sim.hdist}}
}

