% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variance_functions.R
\name{CV}
\alias{CV}
\title{Coefficient of Variation (CV)}
\usage{
CV(x)
}
\arguments{
\item{x}{numeric vector.}
}
\value{
CV estimate for x.
}
\description{
\code{CV} is a function that allows the calculation of the
 \strong{coefficient of variation} for a vector of numbers.
}
\examples{
# Examples
set.seed(1234)
CV(rnorm(100))
# Observed Data (Red Porgy data set)
data(RedPorgy)
CV(x = RedPorgy$R1.Age)
}
\seealso{
Other Error Estimates: \code{\link{RSE}}, \code{\link{SE}}
}

