% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variance_functions.R
\name{describe}
\alias{describe}
\title{Descriptive statistics}
\usage{
describe(x, confidence = 0.05, digits = 2)
}
\arguments{
\item{x}{Numeric vector.}

\item{confidence}{confidence level (default of 0.05) for resulting
confidence interval calculation.}

\item{digits}{integer (defaults to 2) specifying the number of significant
digits for the resulting descriptive statistics}
}
\value{
A numeric vector containig the following information:
   \describe{
     \item{n}{sample size}
     \item{mean}{mean}
     \item{StDev}{standard deviation}
     \item{StErr}{standard error of the mean}
     \item{CI}{confidence interval width}
     \item{RSE}{relative standard error}
     \item{RMSE}{root mean square deviation}
     \item{MAE}{mean absolute error}
     }
}
\description{
\code{describe} constructs a named vector that contains various descriptive statistics of numeric data.
}
\examples{
# Example
data(BSB)
describe(BSB$TL, digits = 4)
# Example 2
with(BSB, aggregate(TL ~ Age, FUN = describe, digits = 4))
}

