% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_correlation_functions.R
\name{dist.max}
\alias{dist.max}
\title{Calculate maximum distance among spatial points}
\usage{
dist.max(data, prop)
}
\arguments{
\item{data}{numeric data frame with two columns.  First column should
represent the "x"-coordinates and the 2nd column represents the
"y"-coordinates}

\item{prop}{numeric variable}
}
\value{
Character string representing a model name which represents the
  proposed model structure.
}
\description{
\code{dist.max} computes the maximum distance among a set of spatial points
}

