% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/precision_functions.R
\name{dist.plot}
\alias{dist.plot}
\title{Distribution Plot}
\usage{
dist.plot(R1, R2, R1.label = "Reader 1", R2.label = "Reader 2")
}
\arguments{
\item{R1}{Numeric vector of independent age determinations from a single
fish from Reader 1.  This must be the same length as R2.}

\item{R2}{Numeric vector of independent age determinations from a single
fish from Reader 2.  This must be the same length as R1.}

\item{R1.label}{Text label for the x-axis (defaults to Reader 1) for the
resultant plot}

\item{R2.label}{Text label for the y-axis (defaults to Reader 2) for the
resultant plot}
}
\value{
A \code{\link{xyplot}} using the package lattice depicting the
   distribution of reader 2 ages, given reader 1 assigned a fish a certain
   age.
}
\description{
\code{dist.plot} creates a distribution plot that helps the researcher
   assess whether bias is present between two sets of age reads on the same
   individual fish.  The plot is constructed using lattice graphics.  It
   provides three unique pieces of information:
   \enumerate{
       \item 1:1 line showing the expected distribution of reads given no
           bias in age determinations between the paired reads
       \item linear regression line fit to the observed paired data for the
           individual reads
       \item text providing the number of times reader 2 assigned a fish
           an age, given reader 1 assigned that fish age "x"
   }
}
\examples{
# Example with Red Porgy data
data(RedPorgy)
dist.plot(R1 = RedPorgy$R3.Age, R2 = RedPorgy$R4.Age, R1.label = "Reader 3",
   R2.label = "Reader 4")
}
\seealso{
Other Bias Plot Functions: 
\code{\link{bias.plot}()}
}
\concept{Bias Plot Functions}
