% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Index_functions.R
\name{GAM.Formulas}
\alias{GAM.Formulas}
\title{Develop formulas for GAM models}
\usage{
GAM.Formulas(Response, names, VarType, basis = "cr")
}
\arguments{
\item{Response}{Character string specifying the name of the response
variable.  This should match a specific column name in your data frame.}

\item{names}{character vector representing the names of unique
co-variates you want to include in your GAM model.  These should match
specific column names in your data frame.}

\item{VarType}{character string denoting the type of co-variate (continuous
("C") or discrete ("F")) for each co-varaite listed in *predictors*.  This
character string must be the same length as the character string specified
by *predictors*.  If a value other than "C" or "F" is provided, the
function will return an error.}

\item{basis}{a two letter character string indicating the (penalized)
smoothing basis to use (e.g., "tp" for thin plate regression spline, "cr"
for cubic regression spline).  see \code{\link[mgcv]{s}} and
\code{\link[mgcv]{smooth.terms}} for more information.}
}
\value{
Character string representing a potential GAM model formulat which
  represents the proposed model structure defined by the function inputs.
}
\description{
\code{GAM.Formulas} generates a character vector that can be used to
  represent formulas for use in individual GAM model runs.  The intent is to
  quickly generate a GAM model formula that incorporates information from
  several potential covariates, the type of covariate utilized, and the
  spline "basis" you specify.
}
\examples{
data(iris)
form <- GAM.Formulas(Response = "Sepal.Length", names =
  c("Petal.Length", "Petal.Width", "Sepal.Width", "Species"), VarType =
  c(rep("C", 3), "F"), basis = "tp")
form
}
\seealso{
Other Model Formulas: 
\code{\link{GLM.Formulas}()},
\code{\link{Model.Formulas}()},
\code{\link{ZI.Formulas}()}
}
\concept{Model Formulas}
