% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Francis_bomb_functions.R
\name{get.h.from.bias}
\alias{get.h.from.bias}
\title{Calculate Relationship Between "h" and Bias}
\usage{
get.h.from.bias(
  ref,
  test,
  hdist,
  bias.step = 5,
  bias.range = NULL,
  ref.x = "x",
  ref.y = "y",
  age = "Fish.Age",
  cy = "Catch.Year",
  age.err = "SE.Age",
  s.age = "Sample.Age",
  test.x = "Sample.Year",
  test.y = "F14C"
)
}
\arguments{
\item{ref, test}{dataframes representing reference and test data sets,
respectively, as described in function \code{\link[FishyR]{calc.bias}}}

\item{hdist}{simulated set of h values as generaed by function
\code{\link[FishyR]{sim.hdist}}}

\item{bias.step, bias.range}{the function calculates \code{h} for each value
of bias in \code{seq(bias.range[1], bias.range[2], bias.step)}.  If bias
.range == NULL, then bias.range is initially set to c(-100, 100) but then
is truncated to be just wide enough to include the 95% confidence
interval for h.}

\item{ref.x, ref.y}{character string giving the name of the x and y-values in
the provided \code{ref} data set, respectively.}

\item{age, age.err, cy, s.age, test.x, test.y}{character strings giving the
names of the columns representing ages, age error estimates, capture
year, sample ages, x-variable, and y-variable in the \code{test}
dataframe.}
}
\value{
A list.
}
\description{
\code{get.h.from.bias} calculates the relationship between \code{h} and
   bias, returning a list with components \code{hval} and
   \code{bias.conf.int}
}
\examples{
data(snapper)
snapper.red <- snapper[snapper$C14year >= 1955 & snapper$C14year <= 1972, ]
data(bluenose)
bluenose.red <- bluenose[bluenose$C14year >= 1955 &
  bluenose$C14year <= 1972, ]
bluenose.red <- bluenose.red[bluenose.red$C14 >= -506 &
  bluenose.red$C14 <= 94.0, ]
ref <- get.ref.line(snapper.red$C14year, snapper.red$C14)
hdist <- sim.hdist(ref = snapper.red, test = bluenose.red, nsim = 5000,
   diag = 0)
get.h.from.bias(ref = ref, test = bluenose.red, hdist = hdist, bias.step = 5,
   bias.range = NULL, age = "age", cy = "catch.year", age.err = "age.err",
   s.age = "samp.age", test.x = "C14Year", test.y = "C14")
}
\seealso{
Other Bomb Radiocarbon Analyses: 
\code{\link{bias.plot.BR}()},
\code{\link{calc.bias}()},
\code{\link{calc.h}()},
\code{\link{closest.pt}()},
\code{\link{get.nearest}()},
\code{\link{get.ref.line}()},
\code{\link{sim.hdist}()}
}
\concept{Bomb Radiocarbon Analyses}
