% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Francis_bomb_functions.R
\name{get.nearest}
\alias{get.nearest}
\title{Find Nearest Point}
\usage{
get.nearest(line, x, y)
}
\arguments{
\item{line}{a list, with components x and y (each an n-vector),
containing the series of n points definign the line}

\item{x, y}{- m-vectors defining m points}
}
\value{
Returns a vecto of lenght m.  This has value NA for any point whose
   y value lies outside the range of the y values in the line.
}
\description{
\code{get.nearest} for each point (x,y), returns the x-coordinate of the
   point on the line that is closest to it in the x-direction.  The lie is
   a linear spline defined by joining a series of poitns by straight lines.
}
\examples{
data(snapper)
snapper.red <- snapper[snapper$C14year >= 1955 & snapper$C14year <= 1972, ]
data(bluenose)
bluenose.red <- bluenose[bluenose$C14year >= 1955 &
  bluenose$C14year <= 1972, ]
bluenose.red <- bluenose.red[bluenose.red$C14 >= -506 &
  bluenose.red$C14 <= 94.0, ]
ref <- get.ref.line(snapper.red$C14year, snapper.red$C14)
get.nearest(ref, bluenose.red$C14year, bluenose.red$C14)
}
\seealso{
Other Bomb Radiocarbon Analyses: \code{\link{calc.bias}},
  \code{\link{calc.h}}, \code{\link{closest.pt}},
  \code{\link{get.h.from.bias}},
  \code{\link{get.ref.line}}, \code{\link{plot.bias}},
  \code{\link{sim.hdist}}
}

