% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Francis_bomb_functions.R
\name{get.ref.line}
\alias{get.ref.line}
\title{Isotonic/Montone Regression Best Fit}
\usage{
get.ref.line(x, y)
}
\arguments{
\item{x}{numeric vector representing the \code{x}-coordinates of the
regression points.}

\item{y}{numeric vector representing the \code{y}-coordinates of the
regression points.}
}
\value{
Object of class \code{\link{list}} containing the results of the
   \code{link{isoreg}} function.  The \code{\link{list}} will be of length
   two:
   \describe{
     \item{x}{original (constructed) abscissa values \code{x}, sorted in
         ascending order}
     \item{y}{fitted values corresponding to \code{ordered x} values}
     }
}
\description{
\code{get.ref.line} generates a list containing \code{\link{isoreg}} function
   model fits to \code{x} and \code{y} data.  The function 1st checks to see
   if multiple \code{y} values are present for a given \code{x}. If there
   are, it determines the \code{\link{mean}} \code{y} for each variable of
   \code{x}" before using the function \code{\link{isoreg}} to compute the
   isotonic (monotonely increasing nonparametric) least squares regression
   which is piecewise constant that  bests fits the data.
}
\examples{
set.seed(1234)
x <- 1:100
y <- x + rnorm(100, 0, 5)
get.ref.line(x, y)
}
\seealso{
\code{\link{isoreg}}

Other Bomb Radiocarbon Analyses: 
\code{\link{bias.plot.BR}()},
\code{\link{calc.bias}()},
\code{\link{calc.h}()},
\code{\link{closest.pt}()},
\code{\link{get.h.from.bias}()},
\code{\link{get.nearest}()},
\code{\link{sim.hdist}()}
}
\concept{Bomb Radiocarbon Analyses}
