% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Index_functions.R
\name{GLM.Formulas}
\alias{GLM.Formulas}
\title{Develop formulas for GLM models}
\usage{
GLM.Formulas(Predictors, Response, offset = NULL, VarType, names)
}
\arguments{
\item{Predictors}{numeric vector representing 'value' for each co-variate
included in the model, where the definition of 'value' is different
depending on the type of covariate in question.  For discrete covariates,
a value of 0 indicates that co-variate should not be included in the
model; all other values indicate the variable will be included.  For
continuous covariates, a value of 0 indicates that co-variate should not
be included in the model; positive values indicate the polynomial order
for that co-variate.}

\item{Response}{Character string specifying the name of the response
variable.  This should match a specific column name in your data frame.}

\item{offset}{character vector representing an "offset" variable to be used
in the formula.}

\item{VarType}{character string denoting the type of co-variate (continuous
("C") or discrete ("F")) for each co-varaite listed in *predictors*.  This
character string must be the same length as the character string specified
by *predictors*.  If a value other than "C" or "F" is provided, the
function will return an error.}

\item{names}{character vector representing the names of unique
co-variates you want to include in your GLM model.  These should match
specific column names in your data frame.}
}
\value{
Character string representing a potential GLM model formula which
  represents the proposed model structure defined by the function inputs.
}
\description{
\code{GLM.Formulas} generates a character vector that can be used to
  represent formulas for use in individual GLM model runs.  The intent is to
  quickly generate a GLM model formula that incorporates information from
  several potential covariates and the type of covariate utilized.
}
\examples{
data(iris)
form <- GLM.Formulas(Response = "Sepal.Length", Predictors = c(1, 2, 0, 3),
names = c("Petal.Length", "Petal.Width", "Sepal.Width", "Species"), VarType =
c(rep("C", 3), "F"))
form
m1 <- glm(as.formula(form), data = iris, family="gaussian")
summary(m1)
plot(m1, all.terms = TRUE, seWithMean = TRUE)
}
\seealso{
Other Model Formulas: 
\code{\link{GAM.Formulas}()},
\code{\link{Model.Formulas}()},
\code{\link{ZI.Formulas}()}
}
\concept{Model Formulas}
