% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Index_functions.R
\name{GLM.Select}
\alias{GLM.Select}
\title{Perform Model Selection Using GLMs}
\usage{
GLM.Select(formula, data, family)
}
\arguments{
\item{formula}{character string representing a potential model formula.}

\item{data}{Data frame containing columns wiht names corresponding to the
values in the formula.}

\item{family}{a description of the error distribution and link function to
be used in the model (see \code{\link{glm}} and \code{\link{family}}.
Outside of those values that can be used in \code{\link{glm}}, you can
also specify it as *nb*.  This tells the model to use the
\code{\link[MASS]{glm.nb}} function available in the package **MASS**.}
}
\value{
vector that contains the \code{\link{AIC}},
  \code{\link[MuMIn]{AICc}}, \code{\link{BIC}} and dispersion parameter
  estimate for a given model structure.
}
\description{
\code{GLM.Select} generates a vector that contains an \code{\link{AIC}}, an
  \code{\link[MuMIn]{AICc}}, a \code{\link{BIC}}, and a dispersion parameter
  estimate.  Intent is to use this function in a loop to extract model
  information criterion for a range of model formulations, facilitating the
  automated selection of *best* model.
}
\seealso{
Other Model Selection: 
\code{\link{ZI.Select}()}
}
\concept{Model Selection}
