% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_sets.R
\docType{data}
\name{haddock}
\alias{haddock}
\title{Adult Haddock (Melanogrammus aeglefinus) Bomb Radiocarbon Data}
\format{
A data frame with 11 rows and 7 variables
\describe{
 \item{age}{integer, age estimate}
 \item{catch.year}{capture year of fish}
 \item{C14year}{integer, estimated birth year based on age and capture year}
 \item{C14}{numeric, C-14 estimate}
 \item{C14.err}{numeric, C-14 estimate error estimate}
 \item{age.err}{numeric, age estimate error estimate}
 \item{samp.age}{integer, age estimate of analyzed sample}}
}
\source{
Campana, S.E.  1997.  Use of radiocarbon from nuclear fallout as a
   dated marker in the otoliths of haddock Melanogrammus aeglefinus.  Mar.
   Ecol. Prog. Ser. 150: 49-56.
}
\usage{
haddock
}
\description{
A dataset containing adult Haddock Bomb Radiocarbon data, as provided in
  Campana 1997 (test data in table 1).
}
\keyword{datasets}
