% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Index_functions.R
\name{Index.Grid}
\alias{Index.Grid}
\title{Develop "grid" over which to predict CPUE from an index model based on
  covariate values}
\usage{
Index.Grid(
  X,
  X.type,
  X.name,
  covariates,
  type,
  bin.length = 17,
  cont.length = 51
)
}
\arguments{
\item{X}{numeric vector representing the primary covariate of interest.  As
such, this is the variable you are wanting to get a predicted "effect" for
over a range of values this variable can take.}

\item{X.type}{character vector defining structure of the \code{X} variable of
interest.  See \code{type} for a description of the possible \code{X.type}s}

\item{X.name}{character string defining the name of the \code{X} variable of
interest}

\item{covariates}{numeric list containing additional covariates included in
the final models being considered}

\item{type}{character vector defining structure of each covariate.  It must
be the same length as the \code{covariates} list.  Each covariate must be
defined as one of four values:
\describe{
  \item{factor}{discrete covariate where data fall in discrete bins}
  \item{mean}{any variable for which you want to use the mean value in the
    model prediction frame.  This would be commonly used to represent an
    offset variable or potentially continous covariates}
  \item{numeric}{continuous covariate, i.e., data fall along a continuous
    scale}
  \item{continuous}{continuous covariate, i.e., data fall along a
    continuous scale}
  \item{discrete}{numeric value representing a discrete value for which you
    want to use in a prediction frame for a given covariete.  For example,
    you may want to choose a specific depth or latitude}
  \item{median}{any variable for which you want to use the median value in
    the model prediction frame.  This would be commonly used to represent
    an offset variable or potentially continous covariates}
  \item{mode}{any variable for which you want to use the modal value in the
    model prediction frame.  This would be commonly used to represent an
    offset variable or potentially continous covariates}
}}

\item{bin.length}{numeric value representing the number of discrete points of
each continuous covariate to use in the prediction frame.  Recommendation
is to use an odd number such that the median of the range of the contiuous
variable is used.  Function used the value to choose n = \code{bin.length}
equally spaced points from the minimum of each continuous covariate to the
maximum of each continuous covariate}

\item{cont.length}{numeric value representing the number of discrete points
of each continuous covariate to use in the prediction frame.
Recommendation is to use an odd number such that the median of the range of
the contiuous variable is used.  Function used the value to choose n =
\code{cont.length} equally spaced points from the minimum of each
continuous covariate to the maximum of each continuous covariate. It is
analogous to \code{bin.length}, just allows for additional flexibility of
having more precision for the primary variable of interest}
}
\value{
list the same length as \code{models} that contains summary
  statistics for each model explored
}
\description{
\code{Index.Grid}
}
\examples{
X <- factor(sample(c(seq(1990, 2015, 1)), size = 10000, replace = TRUE))
df <- data.frame(depth = rnorm(n = 10000, mean = 0, sd = 1), temp = rnorm(
n = 10000, mean = 0, sd = 0.5), lat = rnorm(n = 10000, mean =0, sd = 3))
Index.Grid(X = X, X.type = "factor", X.name = "Year", covariates = df,
type = rep("mean", 3))
Index.Grid(X = X, X.type = "factor", X.name = "Year", covariates = df, type =
rep("numeric", 3), bin.length = 3)
}
\seealso{
Other Model Evaluation: 
\code{\link{Index.Summary}()},
\code{\link{disp}()}
}
\concept{Model Evaluation}
