% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Index_functions.R
\name{Index.Summary}
\alias{Index.Summary}
\title{Summarize results of a standardized relative abundance index}
\usage{
Index.Summary(
  Y,
  X,
  grid,
  models,
  model.name = c("Poisson", "NB", "ZIP", "ZINB", "Nominal")
)
}
\arguments{
\item{Y}{numeric vector representing response variable.  In the case of a
relative abundance index, this would represent your CPUE measure}

\item{X}{numeric vector representing the primary covariate of interest.  As
such, this is the variable you are wanting to get a predicted "effect" for
over a range of values this variable can take.}

\item{grid}{prediction grid.  This most commonly would be created via a call
to \code{Index.Grid} if used in a relative abundance index development
question}

\item{models}{character vector of named models for which you want to get an
index summary for}

\item{model.name}{character vector defining model type.  All values of the
character vector must take one of five defined values:
\describe{
  \item{Poisson}{identifies use of a "Poisson GLM" for the development of
    the model}
  \item{NB}{identifies use of a "Negative Binomial GLM" for the development
    of the model}
  \item{ZIP}{identifies use of a "Zero-Inflated Poisson GLM" for the
    development of the model}
  \item{ZINB}{identifies use of a "Zero-Inflated Negative Binomial GLM" for
    the development of the model}
  \item{Nominal}{identifies the calculation of a nominal index}
}}
}
\value{
list the same length as \code{models} that contains summary
  statistics for each model explored
}
\description{
\code{Index.Summary}
}
\seealso{
Other Model Evaluation: 
\code{\link{Index.Grid}()},
\code{\link{disp}()}
}
\concept{Model Evaluation}
