% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variance_functions.R
\name{mae}
\alias{mae}
\title{Mean Absolute Error}
\usage{
mae(x, expected)
}
\arguments{
\item{x}{numeric vector.}

\item{expected}{expectation, given the model, for each value of \strong{x}.}
}
\value{
Based on the assumed underlying model, the estimated MAE given the
 observed data in \code{x}.
}
\description{
\code{mae} is a function that allows the calculation of the
 \strong{mean absolute error (MAE)} for a vector of numbers.
}
\examples{
# Examples
set.seed(1234)
x <- rnorm(100)
mean(x)
mae(x, mean(x))
# Observed Data (Red Porgy data set)
data(RedPorgy)
x <- na.exclude(RedPorgy[, 1])
mae(x = x, expected = mean(x))
}
\seealso{
Other Model Error Functions: \code{\link{rmsd}}
}

