% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MI_functions.R
\name{MIA.Edge}
\alias{MIA.Edge}
\title{Marginal Increment Analysis Plot - Uses only Categorical Edge Codes}
\usage{
MIA.Edge(x, y, x.label = "Month", y.label = "Proportion")
}
\arguments{
\item{x}{vector representing the grouping variable for the marginal
increment analysis.}

\item{y}{vector representing the edge type for each value of "x". Must be
the same length as "x".}

\item{x.label}{label for the x-axis}

\item{y.label}{label for the y-axis}
}
\value{
A \code{\link{barchart}} using the package lattice depicting the
   proportional edge type distribution by unique "x".
}
\description{
\code{MIA.Edge} creates a marginal increment analysis plot.
}
\examples{
# Example with Sheepshead Data
data(Sheepshead)
Month <- Sheepshead$Month
Edge <- with(Sheepshead, ifelse(MI/Prev.Inc>=0.6667, "Wide", "Narrow"))
MIA.Edge(x = Month, y = Edge)
}
\seealso{
\code{\link{MIA}}
}

