% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MI_functions.R
\name{MIA}
\alias{MIA}
\title{Marginal Increment Analysis Plot}
\usage{
MIA(
  Data,
  Statistics,
  x.label = "Month",
  pch.raw = 20,
  col.raw = "#0080ff",
  alpha.raw = 0.5,
  cex.raw = 0.8,
  pch.fit = 16,
  col.fit = "black",
  cex.fit = 1,
  col.text = "Black",
  cex.text = 1
)
}
\arguments{
\item{Data}{Data frame containing, at a minimum, two columns.  The first
column should represnt the x-variable.  The 2nd column should represent
the y-variable.  Other columns may be present.}

\item{Statistics}{Data frame containing, at a minimum, four colummns.  The
first column should represent the x-variable.  The 2nd column should
represent the y-variable, typically the mean of y at a given x.  The 3rd
column should represent a measure of variability, typically the confidence
interval of y given x.  The 4th column should represent the number of
samples of y at a given level of x.  Other columns may be present in the
data frame.}

\item{x.label}{Text label for the x-axis of the plot}

\item{pch.raw}{symbol to use for the raw x and y data found in \code{Data}}

\item{col.raw}{color to use for the raw x and y data found in \code{Data}}

\item{alpha.raw}{amount of transparency of the raw x and y data found in
\code{Data}}

\item{cex.raw}{symbol size to use for the raw x and y data found in
\code{Data}}

\item{pch.fit}{symbol to use for the summarized marginal increment data
found in \code{Statistics}}

\item{col.fit}{color to use for the summarized marginal increment data
found in \code{Statistics}}

\item{cex.fit}{symbol size to use for the summarized marginal increment
data found in \code{Statistics}}

\item{col.text}{color of text to use for depiciting the sample size at each x
variable level}

\item{cex.text}{text size to use for depiciting the sample size at each x
variable level
bias plot between Reader 2 and Reader 1 ages.}
}
\value{
A \code{\link{xyplot}} using the package lattice depicting the
   results of a marginal increment analysis.
}
\description{
\code{MIA} creates a marginal increment analysis plot.
}
\examples{
# Example with Sheepshead Data
data(Sheepshead)
Data <- data.frame(Month = Sheepshead$Month,
                   Perc.Comp = Sheepshead$MI/Sheepshead$Prev.Inc)
Data$Perc.Comp <- with(Data, ifelse(Perc.Comp >1, 1, Perc.Comp))
Statistics <- with(Data, aggregate(Perc.Comp ~ Month, FUN = describe,
   digits = 4 ))
Statistics <- cbind(Statistics[-ncol(Statistics)],
   Statistics[[ncol(Statistics)]])
MIA(Data = Data, Statistics = Statistics[, c(1, 3, 6, 2)])
}
\seealso{
\code{\link{MIA.Edge}}
}
