% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_correlation_functions.R
\name{Moran.I.calc}
\alias{Moran.I.calc}
\title{Moran's I Test for Spatial Autocorrelation Among Residuals}
\usage{
Moran.I.calc(data)
}
\arguments{
\item{data}{numeric data frame with three columns.  First column should
represent the "x"-coordinates, 2nd column represents the "y"-coordinates,
and the 3rd column represents a numeric vector of model residuals}
}
\value{
A list containing the elements \code{observed}, \code{expected},
  \code{sd}, & \code{p.value}
}
\description{
\code{Moran.I.calc} computes Moran's I autocorrelation coefficient of *resid*
  given a matrix of weights defined as the inverse of a distance matrix of
  spatial coordinates.
}
\seealso{
\code{\link[ape]{Moran.I}}
}
