% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_sets.R
\docType{data}
\name{nwa}
\alias{nwa}
\title{Northwest Atlantic Bomb Radiocarbon Reference Data}
\format{
A data frame with 11 rows and 8 variables
\describe{
 \item{age}{integer, age estimate}
 \item{catch.year}{capture year of fish}
 \item{C14year}{integer, estimated birth year based on age and capture year}
 \item{C14}{numeric, C-14 estimate}
 \item{C14.err}{numeric, C-14 estimate error estimate}
 \item{age.err}{numeric, age estimate error estimate}
 \item{samp.age}{integer, age estimate of analyzed sample}
 \item{YC}{integer, capture year}}
}
\source{
Campana, S.E., Casselman, J.M., and Jones, C.M.  2008.  Bomb
   radiocarbon chronologies in the Arctic, with implications for the age
   validation of lake trout (Salvelinus namaycush) and other Arctic
   species.  Can. J. Fish. Aquat. Sci. 65(4): 733-743.
}
\usage{
nwa
}
\description{
A dataset containing bomb radiocarbon data derived from juvenile Haddock,
   redfish (Sebastes spp.) and yellowtail flounder collected in the
   Northwest Atlntic, adult Haddock Bomb Radiocarbon data, as provided in
   Campana et al. 2008 (Table 2).
}
\keyword{datasets}
