% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/precision_functions.R
\name{perc.agr}
\alias{perc.agr}
\title{Percent Agreement Estimation}
\usage{
perc.agr(Data, levels = 2)
}
\arguments{
\item{Data}{Numeric data frame, with no \code{\link{NA}} values,
representing paired age reads on the same individual fish.}

\item{levels}{Numeric integer representing the maximum number of years you
want to consider for the percent agreement analysis}
}
\value{
Object of class \code{\link{data.table}} and
   \code{\link{data.frame}} containing the percent agreement metrics for
   each year from +/- years to \code{levels}.  The \code{\link{data.table}}
   will contain four columns:
   \describe{
     \item{Scale}{+/- number of years considered for percent agreement
         calculation}
     \item{n}{total sample size for the analysis}
     \item{agree}{# of samples agreed upon by the readers, given that scale}
     \item{perc.agree}{proportinal agreement among the individual age reads}
     }
}
\description{
\code{perc.agr} calculates the percent agreement of two sets of age reads on
   the same fish.  This function will calculate the percent agreement from
   +/- 0 years to +/- the number of years defined by \code{levels},
   by 1 year increments.
}
\examples{
#Example - RedPorgy Dataset
data(RedPorgy)
perc.agr(na.exclude(RedPorgy[, c(1, 2)]), levels = 3)
}

