% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RedDrum.R
\docType{data}
\name{RedDrum}
\alias{RedDrum}
\title{Red Drum catches from the South Carolina Department of Natural Resources
Trammel Net Survey}
\format{

}
\usage{
RedDrum
}
\description{
.
}
\keyword{datasets}
