% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_sets.R
\docType{data}
\name{RedPorgy}
\alias{RedPorgy}
\title{Age Determinations of Red Porgy}
\format{A data frame with 1511 rows and 4 variables:
\describe{
 \item{R1.Age}{biological age, in years, determiend by reader 1}
 \item{R2.Age}{biological age, in years, determiend by reader 2}
 \item{R3.Age}{biological age, in years, determiend by reader 3}
 \item{R4.Age}{biological age, in years, determiend by reader 4}
}}
\source{
South Carolina Deparment of Natural Resources, Coastal Finfish Section
 (ballengerj@dnr.sc.gov)
}
\usage{
RedPorgy
}
\description{
A dataset containing multiple age determinations for Red Porgy made by four
   different readers.  Each reader did not age each individual fish.
}
\keyword{datasets}

