% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variance_functions.R
\name{rmsd}
\alias{rmsd}
\title{Root Mean Square Deviation}
\usage{
rmsd(x, expected)
}
\arguments{
\item{x}{numeric vector.}

\item{expected}{expectation, given the model, for each value of \strong{x}.}
}
\value{
Based on the assumed underlying model, the estimated RMSD given the
 observed data in \code{x}.
}
\description{
\code{rmsd} is a function that allows the calculation of the
 \strong{root mean square deviation (RMSD)} for a vector of numbers.  This
 is also referred to as the '\emph{root mean squared error}' in the
 literature.
}
\examples{
# Examples
# Generated Data
set.seed(1234)
x <- rnorm(100)
mean(x)
rmsd(x, mean(x))
# Observed Data (Red Porgy data set)
data(RedPorgy)
x <- na.exclude(RedPorgy[, 1])
rmsd(x = x, expected = mean(x))
}
\seealso{
Other Model Error Functions: \code{\link{mae}}
}

