% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rounding_functions.R
\name{roundpow}
\alias{roundpow}
\title{Rounding to a Power of 10}
\usage{
roundpow(x, power = 1)
}
\arguments{
\item{x}{a numeric vector.}

\item{power}{positive integer indicating the power of ten to round too.}
}
\value{
numeric vector of \code{x} rounded to the power of 10 specified by
\code{power}.
}
\description{
\code{roundpow} is a function that allows for traditional rounding of
 numbers to a power of 10.  The built in R rounding function
 (\code{\link{round}}) from the R base package follows the standard of the
 IEC 60559 when determining how to round off the number 5. This rule states
 that you '\emph{go to the even digit}'. This is different from how most
 people learn to round off 5 in school (and different from how Excel rounds
 numbers), with one rounding 5 off to the next highest absolute value
 integer. The function \code{\link{roundup}} in this package allows one to
 apply the traditinal rounding measure.  This function extends
 \code{\link{roundup}} by allowing one to round numbers to a power of 10.
}
\examples{
# Examples
roundpow(125, power=1)
roundpow(125, power=2)
roundpow(10555, power=1)
roundpow(10555, power=2)
roundpow(10555, power=3)
roundpow(10555, power=4)
# Comparing IEEE Rounding vs. Traditional Rounding
data(BSB)
IEEE <- with(BSB, round(TL / 10) * 10)
Trad <- with(BSB, roundpow(TL, 1))
plot(Trad ~ jitter(IEEE), xlab = "IEEE Total Length (mm)",
 ylab = "Traditional Total Length (mm)")
}
\seealso{
Other Rounding functions: 
\code{\link{roundup}()}
}
\concept{Rounding functions}
