% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rounding_functions.R
\name{roundup}
\alias{roundup}
\title{Rounding}
\usage{
roundup(x, numdigits = 0)
}
\arguments{
\item{x}{a numeric vector.}

\item{numdigits}{positive integer indicating the number of decimal places to
be used.}
}
\value{
numeric vector of \code{x} rounded to the number of digits specified
 by \code{numdigits}.
}
\description{
\code{roundup} is a function that allows for traditional rounding of numbers.
 The built in R rounding function (\code{\link{round}}) from the R base
 package follows the standard of the IEC 60559 when determining how to round
 off the number 5. This rule states that you '\emph{go to the even digit}'.
 This is different from how most people learn to round off 5 in school (and
 different from how Excel rounds numbers), with one rounding 5 off to the
 next highest absolute value integer. To illustrate the difference, where
 the base function rounds 0.5 and -2.5  to 0 and -2, respectively, roundup
 rounds 0.5 and -2 to  1 and -3, respectively.
}
\examples{
# Examples
roundup(-2.25, numdigits=1)
round(.5 + -2:4) # IEEE rounding
roundup(.5+ -2:4) # Traditional rounding
# Comparing IEEE Rounding vs. Traditional Rounding
data(BSB)
IEEE <- with(BSB, round(TL / 10))
Trad <- with(BSB, roundup(TL / 10))
table(IEEE,Trad)
plot(Trad ~ jitter(IEEE), xlab = "IEEE Total Length (cm)",
 ylab = "Traditional Total Length (cm)")
}
\seealso{
\code{\link{round}} for base rounding function.

Other Rounding functions: \code{\link{roundpow}}
}

