% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variance_functions.R
\name{RSE}
\alias{RSE}
\title{Relative (Percent) Standard Error}
\usage{
RSE(x)
}
\arguments{
\item{x}{numeric vector.}
}
\value{
RSE estimate for x.
}
\description{
\code{RSE} is a function that allows the calculation of the
 \strong{relative standard error (RSE)} for a vector of numbers.  RSE is also known as \strong{Percent Standard Error (PSE)}.
}
\examples{
# Examples
set.seed(1234)
RSE(rnorm(100))
# Observed Data (Red Porgy data set)
data(RedPorgy)
RSE(x = RedPorgy$R1.Age)
}
\seealso{
Other Error Estimates: 
\code{\link{CV}()},
\code{\link{SE}()}
}
\concept{Error Estimates}
