% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variance_functions.R
\name{runmean}
\alias{runmean}
\title{Calculate Smoothed Running Mean}
\usage{
runmean(x, window, align = "center")
}
\arguments{
\item{x}{numeric vector representing a series of observations.}

\item{window}{integer width of the rolling window.}

\item{align}{character specifying whether the index of the result should be
left- or right-aligned or centered (default) compared to the rolling window
of observations.}
}
\value{
An object of the same class as x with the rolling mean with smoothed
 ends.
}
\description{
\code{errbar} is a function that allows the calculation of smoothed running
 means for a numeric vector.  It uses the \code{\link[zoo]{rollmean}}
 function
 from the package \code{\link{zoo}} to compute rolling means for the numeric
 vector and then the \code{\link{smoothEnds}} function from the package
 \code{\link{stats}} to smooth the end points of the resulting vector.
}
\examples{
set.seed(1234)
x <- seq(1, 50, 1)
y <- (2 + 2 * x) * rnorm(50, 1, 0.2)
runmean(x = y, window=3)
}
\seealso{
\code{\link[zoo]{rollmean}}, \code{\link{smoothEnds}}.
}
