% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variance_functions.R
\name{SE}
\alias{SE}
\title{Standard Error}
\usage{
SE(x)
}
\arguments{
\item{x}{numeric vector.}
}
\value{
Standard error estimate for x.
}
\description{
\code{SE} is a function that allows the calculation of the
 \strong{standard error} for a vector of numbers.
}
\examples{
# Examples
set.seed(1234)
SE(rnorm(100))
# Observed Data (Red Porgy data set)
data(RedPorgy)
SE(x = RedPorgy$R1.Age)
}
\seealso{
Other Error Estimates: \code{\link{CV}}, \code{\link{RSE}}
}

