% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_sets.R
\docType{data}
\name{Sheepshead}
\alias{Sheepshead}
\title{Sheepshead Marginal Increment Data}
\format{A data frame with 416 rows and 4 variables
\describe{
 \item{Month}{numeric, month of capture of individual fish}
 \item{MI}{numeric, marginal increment measurement, in micrometers}
 \item{Prev.Inc}{numeric, width of previous increment, in micrometers}
 \item{Oto.Age}{numeric, age, in years, of individual fish}
}}
\source{
Joseph C. Ballenger (\email{ballengerj@dnr.sc.gov})
}
\usage{
Sheepshead
}
\description{
A dataset containing Sheepshead marginal increment measurements.
}
\keyword{datasets}

