% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_sets.R
\docType{data}
\name{snapper}
\alias{snapper}
\title{Snapper (Pagrus auratus) Bomb Radiocarbon Data}
\format{
A data frame with 16 rows and 7 variables
\describe{
 \item{age}{integer, age estimate}
 \item{catch.year}{capture year of fish}
 \item{C14year}{integer, estimated birth year based on age and capture year}
 \item{C14}{numeric, C-14 estimate}
 \item{C14.err}{numeric, C-14 estimate error estimate}
 \item{age.err}{numeric, age estimate error estimate}
 \item{samp.age}{integer, age estimate of analyzed sample}}
}
\source{
Kalish, J.M.  1993.  Pre- and post-bomb radiocarbon in fish
   otoliths.  Earth Planet. Sci. Lett. 114(4): 549-554.
}
\usage{
snapper
}
\description{
A dataset containing Snapper Bomb Radiocarbon data, as provided in Kalish
  1993.
}
\keyword{datasets}
