% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot_functions.R
\name{Stat.Calc}
\alias{Stat.Calc}
\title{Calculate bootstrap statistics}
\usage{
Stat.Calc(X, data, index)
}
\arguments{
\item{X}{TBD}

\item{data}{TBD}

\item{index}{TBD}
}
\value{
numeric vector with four points:
  \describe{
    \item{Points}{}
    \item{Group}{}
    \item{Factor}{}
    \item{Value}{}
  }
}
\description{
\code{Stat.Calc} calculates statistics (i.e., variance and cv estimates)
  about a relative abundance index from a bootstrap analysis.
}
\seealso{
Other Index Bootstrap Functions: 
\code{\link{Boot.Diag}()},
\code{\link{ZI.Index.Boot}()}
}
\concept{Index Bootstrap Functions}
