% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/precision_functions.R
\name{symmetry}
\alias{symmetry}
\title{Bowker's Symmetry Test}
\usage{
symmetry(data)
}
\arguments{
\item{data}{Numeric data frame, with no \code{\link{NA}} values,
representing paired age reads on the same individual fish.}
}
\value{
Object of class \code{\link{data.frame}} containing the results of
   the Bowker's symmetry test.  The \code{\link{data.frame}} will contain
   three columns:
   \describe{
     \item{chisq}{Chi-square statistic}
     \item{df}{estimated degrees of freedom for the chi-square test}
     \item{p.value}{p-value of the symmetry test}
     }
}
\description{
\code{symmetry} Performs Bowker's symmetry test comparing two sets of age
   reads on the same fish.
}
\examples{
#Example - RedPorgy Dataset
data(RedPorgy)
symmetry(na.exclude(RedPorgy[, c(1, 2)]))
}

