% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_correlation_functions.R
\name{Variogram_Plot}
\alias{Variogram_Plot}
\title{Calculate variogram summary statistics and plots}
\usage{
Variogram_Plot(data, mindist, maxdist, length, type, map = FALSE,
  cloud = FALSE, bins = 100)
}
\arguments{
\item{data}{numeric data frame with two columns.  First column should
represent the "x"-coordinates and the 2nd column represents the
"y"-coordinates.  These columns must be titled "X" and "Y", respectively}

\item{mindist}{numeric value representing the minimum distance over which
variogram statistics should be calculated}

\item{maxdist}{numeric value representing the maximum distance over which
variogram statistics should be calculated}

\item{length}{numeric value representing how many points between
\code{mindist} and \code{maxdist} variogram statistics should be calculated}

\item{type}{logical; if \code{TRUE}, use Cressie's robust variogram
estimate; if \code{FALSE} use the classical method of moments variogram
estimate.  See \code{\link[gstat]{variogram}}}

\item{map}{logical; See \code{\link[gstat]{variogram}}}

\item{cloud}{logical; See \code{\link[gstat]{variogram}}}

\item{bins}{numeric value used to calculate the \code{width} (See
\code{\link[gstat]{variogram}}) of subsequent distance intervals}
}
\value{
Data frame with results statistics and a series of variogram plots
  at different threshold distances
}
\description{
\code{Variogram_Plot} is a wrapper funtion that allows the efficient
  computation of numerous variagram statistics and variogram plots over a
  range of considered "plausible" distances.
}

