% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Index_functions.R
\name{ZI.Formulas}
\alias{ZI.Formulas}
\title{Develop formulas for Zero-Inflation models}
\usage{
ZI.Formulas(Predictors, Response, offset = NULL, VarType, names, count.n)
}
\arguments{
\item{Predictors}{numeric vector representing 'value' for each co-variate
included in the model, where the definition of 'value' is different
depending on the type of covariate in question.  For discrete covariates,
a value of 0 indicates that co-variate should not be included in the
model; all other values indicate the variable will be included.  For
continuous covariates, a value of 0 indicates that co-variate should not
be included in the model; positive values indicate the polynomial order
for that co-variate.}

\item{Response}{Character string specifying the name of the response
variable.  This should match a specific column name in your data frame.}

\item{offset}{character vector representing an "offset" variable to be used
in the formula.}

\item{VarType}{character string denoting the type of co-variate (continuous
("C") or discrete ("F")) for each co-varaite listed in *predictors*.  This
character string must be the same length as the character string specified
by *predictors*.  If a value other than "C" or "F" is provided, the
function will return an error.}

\item{names}{character vector representing the names of unique
co-variates you want to include in your zero-inflation model.  These
should match specific column names in your data frame.}

\item{count.n}{integer representing the number of covariates included in the
*count* sub-model of the overall zero-inflation model.  If equal to (or
greater) the length of *names* then all covariates are only included in
the *count* sub-model.  If less than the length of *names*, then those
co-variates at positions greater than *count.n* are included in the
zero-inflation sub-model.}
}
\value{
Character string representing a potential zero-inflation model
  formula which  represents the proposed model structure defined by the
  function inputs.
}
\description{
\code{ZI.Formulas} generates a character vector that can be used to
  represent formulas for use in individual zero-inflation model runs.  The
  intent is to quickly generate a zero-inflation model formula that
  incorporates information from several potential covariates and the type of
  covariate utilized.  The same or different co-variates can be included in
  the different sub-models
}
\examples{
data(iris)
form <- ZI.Formulas(Response = "Sepal.Length", Predictors = c(1, 2, 0, 3, 1,
0, 0, 0), names = rep(c("Petal.Length", "Petal.Width", "Sepal.Width",
"Species"), 2), VarType = rep(c(rep("C", 3), "F"), 2), count.n = 4)
form
}
\seealso{
Other Model Formulas: \code{\link{GAM.Formulas}},
  \code{\link{GLM.Formulas}}, \code{\link{Model.Formulas}}
}

