% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Index_functions.R
\name{ZI.Select}
\alias{ZI.Select}
\title{Perform Model Selection Using Zero-Inflated GLMs}
\usage{
ZI.Select(formula, data, dist, link)
}
\arguments{
\item{formula}{character string representing a potential model formula}

\item{data}{data frame containing columns wiht names corresponding to the
values in the formula}

\item{dist}{see help page for \code{\link[pscl]{zeroinfl}}}

\item{link}{see help page for \code{\link[pscl]{zeroinfl}}}
}
\value{
vector that contains the \code{\link{AIC}},
  \code{\link[MuMIn]{AICc}}, \code{\link{BIC}} and dispersion parameter
  estimate for a given model structure
}
\description{
\code{ZI.Select} generates a vector that contains an \code{\link{AIC}}, an
  \code{\link[MuMIn]{AICc}}, a \code{\link{BIC}}, and a dispersion parameter
  estimate.  Intent is to use this function in a loop to extract model
  information criterion for a range of model formulations, facilitating the
  automated selection of *best* model.
}
\seealso{
Other Model Selection: 
\code{\link{GLM.Select}()}
}
\concept{Model Selection}
